Log error when unable to test serial port connection
Spent a few hours this morning chasing down why I couldn't connect to the AxiDraw V3 -- turns out it was a permission error related to the USB socket my AxiDraw was connected to. This info is captured in the `serial.SerialException`, but that information is unfortunately suppressed by the `pass` statement I've replaced here.